@@ -2,26 +2,25 @@
 using Microsoft.AspNetCore.Mvc;
 using ValidationSample.Shared.Models;
 
-namespace ValidationSample.Controllers
+namespace ValidationSample.Controllers;
+
+[ApiController]
+[Route("[controller]")]
+public class PeopleController : ControllerBase
 {
-    [ApiController]
-    [Route("[controller]")]
-    public class PeopleController : ControllerBase
+    private readonly IValidator<Person> personValidator;
+
+    public PeopleController(IValidator<Person> personValidator)
     {
-        private readonly IValidator<Person> personValidator;
+        this.personValidator = personValidator;
+    }
 
-        public PeopleController(IValidator<Person> personValidator)
-        {
-            this.personValidator = personValidator;
-        }
+    [HttpPost]
+    public IActionResult Save(Person person)
+    {
+        //var validationResults = personValidator.Validate(person);
+        //var isValid = ModelState.IsValid;
 
-        [HttpPost]
-        public IActionResult Save(Person person)
-        {
-            //var validationResults = personValidator.Validate(person);
-            //var isValid = ModelState.IsValid;
-
-            return NoContent();
-        }
+        return NoContent();
     }
 }
